salt/oscarcommerce_base/source/opt/requirements.txt
appdirs
backports.shutil-get-terminal-size
boto
celery
Collectfast
cryptography
decorator
Django
django-contact-form
django-debug-toolbar
django-feedback-form2
django-flatblocks
django-guardian
django-haystack
django-loginas
django-mathfilters
django-mptt
django-mptt-admin
django-ses
django-sitetree
django-storages
django-tinymce4
enum34
gunicorn
ipython
ipython-genutils
librabbitmq
packaging
pathlib2
pexpect
pickleshare
psycopg2
ptyprocess
Pygments
pyparsing
python-dateutil
pytz
requests
setproctitle
simplejson
six
traitlets
#
django-oscar
pycountry
django-compressor